Cambio de codigo a Model.
Hubo una confusión de archivo, lo de repartidorController se pasa a repartidorModel.
